Front end Code:
The front end of this website is built in PHP with HTML and CSS embedded within it. 
The 'Includes' folder:
The file that connects to the database dbh_inc.php
The CSS styling of the forms, tables, and NavBar

The  'Query_pages' Folder:
It has the MySQLi code to query the database by the presented filters or by Catalog Number

The outer Page, Index.php, is the home page.